Pass the last one TestFigure8Unreliable2C
In previous implement, when AppendEntries failed, nextIndex--.
If there are many inconsistent entries(about 500 entries in this case),
it will take too much time to sync.

Pass TestFigure8Unreliable2C via reducing retry AppendEntries times.
1, If follower do not have this PreLog, returns len(rf.log)
2, If PreLog's term is not same as follower's, move to the first one entry of follower's Prelog Term

From raft-extended:
For example, when rejecting an AppendEntries request, the follower can include the term of
the conflicting entry and the first index it stores for that term.

With this information, the leader can decrement nextIndex to bypass all of the conflicting
entries in that term; one AppendEntries RPC will be required for each term with conflicting
entries, rather than one RPC per entry.

In practice, we doubt this optimization is necessary, since failures happen infrequently and
it is unlikely that there will be many inconsistent entries